Bug fixes.
Fixed a bug where players would keep their ability if they activated an ability while one was currently active. Now it will disable the old ability as the new one is activated.

Also fixed a bug where the ball would go too slow or stop. This has been fixed by enforcing a minimum velocity. This can sometimes be a bit glitchy but is rare and better than the ball stopping.