Added events feature with three events.
Added the events feature. Events are like power ups but are activated immediately upon impact and will not favour the person that hit the ball last.

The three events added are: speed (increases the ball's velocity by 50%), grow (increases the ball's size by 50%), and shrink (decreases the ball's size by 25%).